BudgetPlanning/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-jazzmin==2.6.0
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.1
phonenumbers==8.13.19
psycopg2-binary==2.9.5
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
